- Moved quotes module to grouped folder.
