+++ src/app/admin/(modules)/quotes/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from "next";
 import { BreadcrumbDataSetter } from "@/app/admin/_components/Breadcrumbs";
-import QuoteFormLoading from "@/app/admin/quotes/_components/Loading/QuoteFormLoading";
-import CreateQuoteWrapper from "@/app/admin/quotes/_components/CreateQuoteWrapper";
+import QuoteFormLoading from "@/app/admin/(modules)/quotes/_components/Loading/QuoteFormLoading";
+import CreateQuoteWrapper from "@/app/admin/(modules)/quotes/_components/CreateQuoteWrapper";
 import { Fragment, Suspense } from "react";
 
 export const metadata: Metadata = {
